refactor(tests): C9-05/06 split test_tools_recipe.py and test_tools_integrations.py
Split test_tools_recipe.py (1,415 lines) into:
- test_tools_list_recipes.py: list_recipes handler tests
- test_tools_load_recipe.py: load_recipe + migrate_recipe tests
- test_tools_recipe.py: validate_recipe + docstring semantic tests

Move github unit tests (report_bug, _extract_block, _parse_*_result)
from test_tools_integrations.py to test_tools_github.py.
Promote _skill_ok/_skill_fail helpers to tests/server/conftest.py.
